Returning to the old working version.
After a number of attempts,
VAE could not be trained

Many attempts were spent on training U-net AE
for images of 256 pixels, there were different
models, but an acceptable result could
not be achieved...

There were attempts to reduce the horrific size
of this version, insufficiently good models
were obtained. Since there is not enough time,
it was decided to return everything as it was...

Callbacks were also added for logging in Tensorboard.